feat: TNL-10106 Extract course enrollment from student model
- add required import statements to support new file
- redirect external imports from student to course_enrollment, as required
- remove __init__ import statements (not needed)
