+++ common/djangoapps/student/management/commands/bulk_change_enrollment_csv.py
@@ -14,8 +14,8 @@ from django.db import transaction
 from opaque_keys import InvalidKeyError
 from opaque_keys.edx.keys import CourseKey
 
-from common.djangoapps.student.models.student import CourseEnrollment, CourseEnrollmentAttribute, User
-
+from common.djangoapps.student.models.course_enrollment import CourseEnrollment, CourseEnrollmentAttribute
+from common.djangoapps.student.models.student import User
 from common.djangoapps.student.models.student import BulkChangeEnrollmentConfiguration
 
 logger = logging.getLogger('common.djangoapps.student.management.commands.bulk_change_enrollment_csv')
