+++ common/djangoapps/student/management/commands/bulk_unenroll.py
@@ -7,7 +7,7 @@ import unicodecsv
 from django.core.management.base import BaseCommand
 from opaque_keys import InvalidKeyError
 from opaque_keys.edx.keys import CourseKey
-from common.djangoapps.student.models.student import CourseEnrollment, BulkUnenrollConfiguration
+from common.djangoapps.student.models.course_enrollment import CourseEnrollment, BulkUnenrollConfiguration
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
