+++ common/djangoapps/student/management/commands/change_eligibility_deadline.py
@@ -10,7 +10,9 @@ from opaque_keys.edx.keys import CourseKey
 
 from common.djangoapps.course_modes.models import CourseMode
 from openedx.core.djangoapps.credit.models import CreditEligibility
-from common.djangoapps.student.models.student import CourseEnrollment, User
+
+from common.djangoapps.student.models.course_enrollment import CourseEnrollment
+from common.djangoapps.student.models.student import User
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
