+++ common/djangoapps/student/management/commands/change_enrollment.py
@@ -9,7 +9,8 @@ from opaque_keys import InvalidKeyError
 from opaque_keys.edx.keys import CourseKey
 
 from openedx.core.djangoapps.credit.email_utils import get_credit_provider_attribute_values
-from common.djangoapps.student.models.student import CourseEnrollment, CourseEnrollmentAttribute, User
+from common.djangoapps.student.models.course_enrollment import CourseEnrollment, CourseEnrollmentAttribute
+from common.djangoapps.student.models.student import User
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
