+++ common/djangoapps/student/management/commands/transfer_students.py
@@ -9,7 +9,7 @@ from django.contrib.auth.models import User  # lint-amnesty, pylint: disable=imp
 from django.db import transaction
 from opaque_keys.edx.keys import CourseKey
 
-from common.djangoapps.student.models.student import CourseEnrollment
+from common.djangoapps.student.models.course_enrollment import CourseEnrollment
 from common.djangoapps.track.management.tracked_command import TrackedCommand
