+++ common/djangoapps/student/management/tests/test_recover_account.py
@@ -16,7 +16,7 @@ from django.test import TestCase, RequestFactory, override_settings
 
 from testfixtures import LogCapture
 from common.djangoapps.student.tests.factories import UserFactory
-from common.djangoapps.student.models.student import AccountRecoveryConfiguration
+from common.djangoapps.student.models.course_enrollment import AccountRecoveryConfiguration
 
 from openedx.core.djangolib.testing.utils import skip_unless_lms
