+++ common/djangoapps/student/models/__init__.py
@@ -1,2 +1,1 @@
-"""Support for student model spread over multiple modules (otherwise would be in single models.py module)"""
-from .student import *
+
