+++ common/djangoapps/student/tests/test_filters.py
@@ -10,7 +10,7 @@ from rest_framework import status
 from xmodule.modulestore.tests.django_utils import ModuleStoreTestCase
 from xmodule.modulestore.tests.factories import CourseFactory
 
-from common.djangoapps.student.models.student import CourseEnrollment, EnrollmentNotAllowed, UnenrollmentNotAllowed
+from common.djangoapps.student.models.course_enrollment import CourseEnrollment, EnrollmentNotAllowed, UnenrollmentNotAllowed
 from common.djangoapps.student.tests.factories import UserFactory, UserProfileFactory
 from openedx.core.djangolib.testing.utils import skip_unless_lms
