+++ lms/urls.py
@@ -13,7 +13,7 @@ from django.views.generic.base import RedirectView
 from edx_api_doc_tools import make_docs_urls
 from edx_django_utils.plugins import get_plugin_url_patterns
 
-from common.djangoapps.student import views as student_views
+from common.djangoapps.student.models.student import views as student_views
 from common.djangoapps.util import views as util_views
 from lms.djangoapps.branding import views as branding_views
 from lms.djangoapps.courseware.masquerade import MasqueradeView
