+++ openedx/core/djangoapps/user_api/accounts/api.py
@@ -13,7 +13,7 @@ from django.core.validators import ValidationError, validate_email
 from django.utils.translation import override as override_language
 from django.utils.translation import gettext as _
 from pytz import UTC
-from common.djangoapps.student import views as student_views
+from common.djangoapps.student.models.student import views as student_views
 from common.djangoapps.student.models import (
     AccountRecovery,
     User,
